Make full size table view as a separate class
